requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
langsmith>=0.0.69
openai>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
jupyter>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
chromadb>=0.4.22
pydantic>=2.5.0
tiktoken>=0.5.1

# The following dependencies were used for the API interface
# and are now optional. Uncomment if you need to re-enable the API.
# fastapi>=0.105.0
# uvicorn>=0.24.0
# streamlit>=1.29.0 